+++ src/main/java/pl/edu/wat/wcy/isi/app/dto/ApproximationPropertiesDTO.java
@@ -4,15 +4,16 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 @Getter
 @Setter
 @Builder
 public class ApproximationPropertiesDTO {
-    private long id;
-    private long userId;
-    private long dataSeriesFileId;
+    private BigInteger id;
+    private BigInteger userId;
+    private BigInteger dataSeriesFileId;
     private int degree;
     private DataSeriesFileDTO dataSeriesFileDTO;
     private Timestamp dateCreate;
