+++ src/main/java/pl/edu/wat/wcy/isi/app/dto/RoleUserDTO.java
@@ -2,13 +2,15 @@ package pl.edu.wat.wcy.isi.app.dto;
 
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class RoleUserDTO {
-    private long id;
+    private BigInteger id;
     private String code;
     private String name;
 }
