+++ src/main/java/pl/edu/wat/wcy/isi/app/repository/RoleUserToUserRepository.java
@@ -5,10 +5,11 @@ import org.springframework.stereotype.Repository;
 import pl.edu.wat.wcy.isi.app.model.entityModels.RoleUserToUserEntity;
 import pl.edu.wat.wcy.isi.app.model.entityModels.UserEntity;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface RoleUserToUserRepository extends JpaRepository<RoleUserToUserEntity, Long> {
+public interface RoleUserToUserRepository extends JpaRepository<RoleUserToUserEntity, BigInteger> {
     List<RoleUserToUserEntity> findByUserByUserId(UserEntity userEntity);
 
     void deleteByUserByUserId(UserEntity userEntity);
